Fixed overflow on calculation

+++ src/gravity.rs
@@ -58,8 +58,8 @@ impl Gravity {
     pub(crate) fn calc_geometry(&self, bounds: &Rectangle, geom: &mut Rectangle) {
         geom.x = bounds.x + (bounds.width as i16 * self.geom.x / 100);
         geom.y = bounds.y + (bounds.height as i16 * self.geom.y / 100);
-        geom.width = bounds.width * self.geom.width / 100;
-        geom.height = bounds.height * self.geom.height / 100;
+        geom.width = (bounds.width as u32 * self.geom.width as u32 / 100) as u16;
+        geom.height = (bounds.height as u32 * self.geom.height as u32 / 100) as u16;
     }
 }
